+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Bump CLI from v2.7.0 to v2.8.1 ([#173](https://github.com/getsentry/sentry-fastlane-plugin/pull/173), [#175](https://github.com/getsentry/sentry-fastlane-plugin/pull/175))
   - [changelog](https://github.com/getsentry/sentry-cli/blob/master/CHANGELOG.md#281)
   - [diff](https://github.com/getsentry/sentry-cli/compare/2.7.0...2.8.1)
+- Support multiple paths in `upload_dif` command (#177)
 
 ## 1.14.0